1.add CashRebate.java.... 2.fix checkOut did not show the decimal point

+++ src/com/htb/cnk/data/OrderedDish.java
@@ -5,7 +5,7 @@ public class OrderedDish {
 	//private final MyOrder OrderedDish;
 	Dish dish;
 	float padQuantity;
-	int phoneQuantity;
+	float phoneQuantity;
 	int status;
 	int tableId;
 	boolean []flavor;
